switch formulaic add* code to use cog
use cog.py to generate the code which is just simple math formulas run
through a basic template. this is more explicit (because you can see
the Python code) in how the values were derived. previously they were
the output of emacs lisp that was cooked in to the source code.